Brings the 0.14.18 release metadata and the NRE/logic fixes that were cut
on the release branch but never merged back. Preserves main's Unity
6000.4/6000.5 compat shims and the AnimateToTransformNode OnInvalid output.

+++ Runtime/Scripts/Expressions/ExpressionFunctions.cs
@@ -636,7 +636,7 @@ namespace Dash
             if (paramType == typeof(double))
             {
                 p_args.HasResult = true;
-                p_args.Result = -(float)evalParams[0];
+                p_args.Result = -(double)evalParams[0];
                 return true;
             }
